+++ turing-react/src/app/console/sn/multi-language/sn.site.multi.language.page.tsx
@@ -1,4 +1,5 @@
 import { ROUTES } from "@/app/routes.const";
+import { LoadProvider } from "@/components/loading-provider";
 import { SNSiteLocaleForm } from "@/components/sn/locales/sn.site.locale.form";
 import { SubPageHeader } from "@/components/sub.page.header";
 import type { TurSNSiteLocale } from "@/models/sn/sn-site-locale.model";
@@ -13,16 +14,20 @@ const turSNSiteLocaleService = new TurSNSiteLocaleService();
 export default function SNSiteMultiLanguagePage() {
   const navigate = useNavigate();
   const { id, localeId } = useParams() as { id: string, localeId: string };
-  const [snLocale, setSnLocale] = useState<TurSNSiteLocale>({} as TurSNSiteLocale);
+  const [snLocale, setSnLocale] = useState<TurSNSiteLocale>();
   const [isNew, setIsNew] = useState<boolean>(true);
   const [open, setOpen] = useState(false);
+  const [error, setError] = useState<string | null>(null);
   useEffect(() => {
-    if (localeId !== "new") {
-      turSNSiteLocaleService.get(id, localeId).then(setSnLocale);
+    if (localeId === "new") {
+      turSNSiteLocaleService.query(id).then(() => setSnLocale({} as TurSNSiteLocale)).catch(() => setError("Connection error or timeout while fetching Multi Language data."));
+    } else {
+      turSNSiteLocaleService.get(id, localeId).then(setSnLocale).catch(() => setError("Connection error or timeout while fetching locale details."));
       setIsNew(false);
     }
   }, [id, localeId])
   async function onDelete() {
+    if (!snLocale) return;
     try {
       if (await turSNSiteLocaleService.delete(snLocale)) {
         toast.success(`The ${snLocale.language} Language was deleted`);
@@ -38,9 +43,9 @@ export default function SNSiteMultiLanguagePage() {
     setOpen(false);
   }
   return (
-    <>
+    <LoadProvider checkIsNotUndefined={snLocale} error={error} tryAgainUrl={`${ROUTES.SN_INSTANCE}/${id}/locale`}>
       {isNew && <SubPageHeader icon={IconLanguage} name="Language" feature="Language" description="Define Multi Languages." />}
-      {!isNew && <SubPageHeader
+      {!isNew && snLocale && <SubPageHeader
         icon={IconLanguage}
         name={snLocale.language}
         feature="Language"
@@ -48,11 +53,11 @@ export default function SNSiteMultiLanguagePage() {
         onDelete={onDelete}
         open={open}
         setOpen={setOpen} />}
-      <SNSiteLocaleForm
+      {snLocale && <SNSiteLocaleForm
         snSiteId={id}
         snLocale={snLocale}
         isNew={isNew}
-      />
-    </>
+      />}
+    </LoadProvider>
   )
 }
