+++ turing-react/src/app/console/sn/result-ranking/sn.site.result.ranking.page.tsx
@@ -1,4 +1,5 @@
 import { ROUTES } from "@/app/routes.const";
+import { LoadProvider } from "@/components/loading-provider";
 import { SNSiteResultRankingForm } from "@/components/sn/result-ranking/sn.site.result.ranking.form";
 import { SubPageHeader } from "@/components/sub.page.header";
 import type { TurSNRankingExpression } from "@/models/sn/sn-ranking-expression.model";
@@ -13,17 +14,20 @@ const turSNRankingExpressionService = new TurSNRankingExpressionService();
 export default function SNSiteResultRankingPage() {
     const navigate = useNavigate();
     const { id, resultRankingId } = useParams() as { id: string, resultRankingId: string };
-    const [resultRanking, setResultRanking] = useState<TurSNRankingExpression>({} as TurSNRankingExpression);
+    const [resultRanking, setResultRanking] = useState<TurSNRankingExpression>();
     const [isNew, setIsNew] = useState<boolean>(true);
     const [open, setOpen] = useState(false);
+    const [error, setError] = useState<string | null>(null);
     useEffect(() => {
-
-        if (resultRankingId !== "new") {
-            turSNRankingExpressionService.get(id, resultRankingId).then(setResultRanking);
+        if (resultRankingId === "new") {
+            turSNRankingExpressionService.query(id).then(() => setResultRanking({} as TurSNRankingExpression)).catch(() => setError("Connection error or timeout while fetching result rankings."));
+        } else {
+            turSNRankingExpressionService.get(id, resultRankingId).then(setResultRanking).catch(() => setError("Connection error or timeout while fetching result ranking details."));
             setIsNew(false);
         }
-    }, [id])
+    }, [resultRankingId])
     async function onDelete() {
+        if (!resultRanking) return;
         try {
             if (await turSNRankingExpressionService.delete(id, resultRanking)) {
                 toast.success(`The ${resultRanking.name} Result Ranking Instance was deleted`);
@@ -39,13 +43,17 @@ export default function SNSiteResultRankingPage() {
         setOpen(false);
     }
     return (
-        <>
-            <SubPageHeader icon={IconNumber123} name={resultRanking.name} feature="Result Ranking"
-                description="Define content that will be featured in the term-based search."
-                onDelete={onDelete}
-                open={open}
-                setOpen={setOpen} />
-            <SNSiteResultRankingForm snSiteId={id} value={resultRanking} isNew={isNew} />
-        </>
+        <LoadProvider checkIsNotUndefined={resultRanking} error={error} tryAgainUrl={`${ROUTES.SN_INSTANCE}/${id}/result-ranking`}>
+            {resultRanking && (
+                <>
+                    <SubPageHeader icon={IconNumber123} name={resultRanking.name} feature="Result Ranking"
+                        description="Define content that will be featured in the term-based search."
+                        onDelete={onDelete}
+                        open={open}
+                        setOpen={setOpen} />
+                    <SNSiteResultRankingForm snSiteId={id} value={resultRanking} isNew={isNew} />
+                </>
+            )}
+        </LoadProvider>
     )
 }
